% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce-df}
\alias{coerce-df}
\alias{.as_df}
\alias{.restore_df_subclass}
\title{Coerce and store \verb{<data.frame>} subclass to \verb{<data.frame>} and
restore \verb{<data.frame>} subclass to \verb{<data.frame>} from attribute.}
\usage{
.as_df(x)

.restore_df_subclass(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A \verb{<data.frame>} or subclass of \verb{<data.frame>}.
}
\description{
Coerce and store \verb{<data.frame>} subclass to \verb{<data.frame>} and
restore \verb{<data.frame>} subclass to \verb{<data.frame>} from attribute.
}
\keyword{internal}
