% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create a list of configuration settings for some details of \code{\link[=sim_linelist]{sim_linelist()}}}
\usage{
create_config(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named elements to replace
default settings. Only if names match exactly are elements replaced,
otherwise the function errors.}
}
\value{
A list of settings for \code{\link[=sim_linelist]{sim_linelist()}}
}
\description{
Create a list of configuration settings for some details of \code{\link[=sim_linelist]{sim_linelist()}}
}
\details{
The \code{config} argument in \code{\link[=sim_linelist]{sim_linelist()}} controls the small details
around time windows around infections (time of first contact and last
contact with infector), and the distribution of the Ct value for confirmed
cases.

Accepted arguments and their defaults are:
\itemize{
\item \code{last_contact_distribution = "pois"}
\item \code{last_contact_distribution_params = c(lambda = 3)}
\item \code{first_contact_distribution = "pois"}
\item \code{first_contact_distribution_params = c(lambda = 3)}
\item \code{ct_distribution = "norm"}
\item \code{ct_distribution_params = c(mean = 25, sd = 2)}
}

These parameters do not warrant their own arguments in
\code{\link[=sim_linelist]{sim_linelist()}} as they rarely need to be changed from their default
setting. Therefore it is not worth increasing the number of \code{\link[=sim_linelist]{sim_linelist()}}
arguments to accommodate these and the \code{config} argument keeps the function
signature simpler and more readable.
}
