% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{.add_date}
\alias{.add_date}
\alias{.add_date_first_contact}
\alias{.add_date_last_contact}
\alias{.add_hospitalisation}
\alias{.add_deaths}
\title{Add event date as column to infectious history \verb{<data.frame>}}
\usage{
.add_date_first_contact(.data, distribution = c("pois", "geom"), ...)

.add_date_last_contact(
  .data,
  outbreak_start_date,
  distribution = c("pois"),
  ...
)

.add_hospitalisation(.data, onset_to_hosp)

.add_deaths(.data, onset_to_death)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation}

\item{distribution}{A \code{character} with the name of the distribution,
following the base R convention for distribution naming (e.g. Poisson
is \code{pois}).}

\item{...}{\link{dots} Extra arguments to be passed to the distribution function
given in the \code{distribution} argument.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{onset_to_hosp}{An \verb{<epidist>} object for the onset to hospitalisation
delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object for the onset to death delay
distribution.}
}
\value{
A \verb{<data.frame>} with one more column than input into \code{.data}.
Unless the column heading is already present in which the data is
overwritten.
}
\description{
The event date could be first contact, last contact or other.
}
\keyword{internal}
