% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{.add_info}
\alias{.add_info}
\alias{.add_names}
\alias{.add_ct}
\title{Add line list information as column to infectious history \verb{<data.frame>}}
\usage{
.add_names(.data)

.add_ct(.data, distribution = c("norm", "lnorm"), ...)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation}

\item{distribution}{A \code{character} with the name of the distribution,
following the base R convention for distribution naming (e.g. Poisson
is \code{pois}).}

\item{...}{\link{dots} Extra arguments to be passed to the distribution function
given in the \code{distribution} argument.}

\item{onset_to_hosp}{An \verb{<epidist>} object with the onset to hospitalisation
delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object with the onset to death delay
distribution.}
}
\value{
A \verb{<data.frame>} with one more column than input into \code{.data}.
Unless the column heading is already present in which the data is
overwritten.
}
\description{
Add line list information as column to infectious history \verb{<data.frame>}
}
\keyword{internal}
