% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.anonymise}
\alias{.anonymise}
\title{Anonymise names}
\usage{
.anonymise(x, string_len = 10)
}
\arguments{
\item{x}{A vector of \code{character} strings.}

\item{string_len}{A single \code{numeric} specifying the number of alphanumeric
characters to use for each anonymising \code{character} string.
Default is \code{10}.}
}
\value{
A vector of \code{character} strings of equal length to the input.
}
\description{
A simple algorithm to replace names with an alphanumeric string with an
fixed number of characters (i.e. \code{\link[=nchar]{nchar()}}) specified by \code{string_len}.
}
\keyword{internal}
