% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{.check_age_df}
\alias{.check_age_df}
\title{Check if \verb{<data.frame>} defining age structure of population is correct}
\usage{
.check_age_df(x)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}
}
\value{
A \verb{<data.frame>}, also called for error side-effects when input is
invalid.
}
\description{
Check if \verb{<data.frame>} defining age structure of population is correct
}
\keyword{internal}
