% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{.check_contact_matrix}
\alias{.check_contact_matrix}
\title{Check if \code{matrix} defining the age-structured contact patterns is correct}
\usage{
.check_contact_matrix(x, population_age)
}
\arguments{
\item{x}{A \code{matrix}.}
}
\value{
A \code{matrix} defining the age-structured contact patterns, also called
for side-effects when input is invalid.
}
\description{
Check if \code{matrix} defining the age-structured contact patterns is correct
}
\keyword{internal}
