% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{.check_linelist}
\alias{.check_linelist}
\title{Check if \R object is line list from \code{\link[=sim_linelist]{sim_linelist()}}}
\usage{
.check_linelist(linelist)
}
\arguments{
\item{linelist}{Line list \verb{<data.frame>} output from \code{\link[=sim_linelist]{sim_linelist()}}.}
}
\value{
Invisibly return the \code{linelist} \verb{<data.frame>}. The function is
called for its side-effects, which will error if the input is invalid.
}
\description{
Check if \R object is line list from \code{\link[=sim_linelist]{sim_linelist()}}
}
\details{
This is a check that the object supplied to \code{linelist} is from the
\code{\link[=sim_linelist]{sim_linelist()}} or \code{\link[=sim_outbreak]{sim_outbreak()}} functions, it is not related to
the class of the object, in other words, it does not check the object is
of class \verb{<linelist>}.
}
\keyword{internal}
