% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{.check_risk}
\alias{.check_risk}
\title{Check if hospitalisation and death risks are correct}
\usage{
.check_risk(x, age_range)
}
\arguments{
\item{x}{A named list of risks}

\item{age_range}{A \code{numeric} vector with two elements.}
}
\value{
A named list of either \code{numeric} or \verb{<data.frame>} risks.
}
\description{
Check if hospitalisation and death risks are correct
}
\keyword{internal}
