% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{.check_risk_df}
\alias{.check_risk_df}
\title{Check if \verb{<data.frame>} defining age-stratified hospitalisation or death risk
is correct}
\usage{
.check_risk_df(x, age_range)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}
}
\value{
A \verb{<data.frame>}, also called for error side-effects when input is
invalid.
}
\description{
Check if \verb{<data.frame>} defining age-stratified hospitalisation or death risk
is correct
}
\keyword{internal}
