% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contacts.R
\name{.create_contacts}
\alias{.create_contacts}
\title{Create contacts table}
\usage{
.create_contacts(
  .data,
  outbreak_start_date,
  contact_distribution,
  age_range,
  contact_tracing_status_probs,
  config
)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{contact_distribution}{An \verb{<epidist>} object for the contact
distribution. This is similar to the offspring distribution, but describes
the heterogeneity of all contact patterns, including contacts that go on
to be infected or not infected.}

\item{age_range}{A \code{numeric} vector with two elements. The first is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]).}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A contacts \verb{<data.frame>}
}
\description{
Create contacts table
}
\keyword{internal}
