% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_contact_matrix}
\alias{.format_contact_matrix}
\title{Format a contact matrix into the \pkg{simulist} style and convert to
\verb{<data.frame>}}
\usage{
.format_contact_matrix(x, population_age)
}
\arguments{
\item{x}{A contact \code{matrix}.}
}
\value{
A contact matrix \verb{<data.frame>}.
}
\description{
Format a contact matrix into the \pkg{simulist} style and convert to
\verb{<data.frame>}
}
\keyword{internal}
