% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sample_age}
\alias{.sample_age}
\title{Sample age of individuals}
\usage{
.sample_age(index_age, population_age, n, contact_df)
}
\arguments{
\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{n}{A single \code{numeric} specifying how many times to sample.}
}
\value{
A \code{numeric} vector.
}
\description{
The ages are sampled either uniformly (using \code{\link[=sample]{sample()}}), or are sampled
with age-structured populations when the \code{population_age} argument is a
\verb{<data.frame>}.
}
\keyword{internal}
