% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_na}
\alias{is_na}
\title{Check if \R object is a single \code{NA}}
\usage{
is_na(x)
}
\arguments{
\item{x}{An \R object}
}
\value{
A single boolean \code{logical}.
}
\description{
Check if an \R object is specifically a single logical \code{\link{NA}} (i.e.
non-vectorised). \code{\link{NA_real_}}, \code{\link{NA_character_}}, \code{\link{NA_integer}},
\code{\link{NA_complex_}} return \code{FALSE}.
}
\keyword{internal}
