% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.R
\name{messy}
\alias{messy}
\title{Create messy line list data}
\usage{
messy(linelist, ...)
}
\arguments{
\item{linelist}{Line list \verb{<data.frame>} output from \code{\link[=sim_linelist]{sim_linelist()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named elements to replace
default settings. Only if names match exactly are elements replaced,
otherwise the function errors.}
}
\value{
A messy line list \verb{<data.frame>}.
}
\description{
Create messy line list data
}
\details{
By default \code{messy()}:
\itemize{
\item Makes 10\% of values missing, i.e. converts to \code{NA}.
\item Introduces spelling mistakes in 10\% of \code{character} columns.
\item Introduce inconsistency in the reporting of \verb{$sex}.
}

To change the defaults of \code{messy()} arguments can be supplied to \code{...}.
Accepted arguments and their defaults are:
\itemize{
\item \code{prop_missing = 0.1}
\item \code{missing_value = NA}
\item \code{prop_spelling_mistakes = 0.1}
\item \code{inconsistent_sex = TRUE}
\item \code{sex_as_numeric = FALSE}
}

When setting \code{sex_as_numeric} to \code{TRUE}, male is set to \code{0} and female
to \code{1}. Only one of \code{inconsistent_sex} or \code{sex_as_numeric} can be \code{TRUE},
otherwise the function will error.
}
\examples{
linelist <- sim_linelist()
messy_linelist <- messy(linelist)

# increasing proportion of missingness to 30\% with a missing value of -99
messy_linelist <- messy(linelist, prop_missing = 0.3, missing_value = -99)

# increasing proportion of spelling mistakes to 50\%
messy_linelist <- messy(linelist, prop_spelling_mistakes = 0.5)

# encode `$sex` as `numeric`
messy_linelist <- messy(
  linelist,
  sex_as_numeric = TRUE,
  inconsistent_sex = FALSE
)
}
