% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulist-package.R
\docType{package}
\name{simulist-package}
\alias{simulist}
\alias{simulist-package}
\title{simulist: Tools to Simulated Line list Data}
\description{
Provide tools to simulate raw case data for an epidemic in the form of line lists, using model outputs, such as compartmental models, branching processes, or network models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/simulist}
  \item \url{https://epiverse-trace.github.io/simulist/}
  \item Report bugs at \url{https://github.com/epiverse-trace/simulist/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Carmen Tamayo \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Pratik R. Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [contributor]
}

}
\keyword{internal}
