% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{.add_date}
\alias{.add_date}
\alias{.add_date_first_contact}
\alias{.add_date_last_contact}
\alias{.add_hospitalisation}
\alias{.add_deaths}
\title{Add event date as column to infectious history \verb{<data.frame>}}
\usage{
.add_date_first_contact(.data, distribution = "pois", ...)

.add_date_last_contact(.data, outbreak_start_date, distribution = "pois", ...)

.add_hospitalisation(.data, onset_to_hosp, hosp_rate)

.add_deaths(.data, onset_to_death, hosp_death_rate, non_hosp_death_rate)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation}

\item{distribution}{A \code{character} with the name of the distribution,
following the base R convention for distribution naming (e.g. Poisson
is \code{pois}).}

\item{...}{\link{dots} Extra arguments to be passed to the distribution function
given in the \code{distribution} argument.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{onset_to_hosp}{An \verb{<epidist>} object for the onset to hospitalisation
delay distribution.}

\item{hosp_rate}{Either a single \code{numeric} for the hospitalisation rate of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation rates. Default is 20\% hospitalisation (\code{0.2}) for the entire
population. See details and examples for more information.}

\item{onset_to_death}{An \verb{<epidist>} object for the onset to death delay
distribution.}

\item{hosp_death_rate}{Either a single \code{numeric} for the death rate for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death rates. Default is 50\% death rate in hospitals
(\code{0.5}) for the entire population. See details and examples for more
information.}

\item{non_hosp_death_rate}{Either a single \code{numeric} for the death rate for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death rates outside of hospitals. Default is 5\% death rate outside
of hospitals  (\code{0.05}) for the entire population. See details and examples
for more information.}
}
\value{
A \verb{<data.frame>} with one more column than input into \code{.data}.
Unless the column heading is already present in which the data is
overwritten.
}
\description{
The event date could be first contact, last contact or other.
}
\keyword{internal}
