% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{.add_date}
\alias{.add_date}
\alias{.add_date_contact}
\alias{.add_hospitalisation}
\alias{.add_outcome}
\title{Add event date as column to infectious history \verb{<data.frame>}}
\usage{
.add_date_contact(
  .data,
  contact_type = c("first", "last"),
  distribution = c("pois", "geom"),
  ...,
  outbreak_start_date = NULL
)

.add_hospitalisation(.data, onset_to_hosp, hosp_risk)

.add_outcome(
  .data,
  onset_to_death,
  onset_to_recovery,
  hosp_death_risk,
  non_hosp_death_risk,
  config
)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation.}

\item{contact_type}{A \code{character} with the type of contact, either first
contact (\code{"first"}), or last contact (\code{"last"}).}

\item{distribution}{A \code{character} with the name of the distribution,
following the base R convention for distribution naming (e.g. Poisson
is \code{pois}).}

\item{...}{\link{dots} Extra arguments to be passed to the distribution function
given in the \code{distribution} argument.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{onset_to_hosp}{An \verb{<epidist>} object, an anonymous function for
the onset to hospitalisation delay distribution, or \code{NA} to not simulate
hospitalisation (admission) dates.}

\item{hosp_risk}{Either a single \code{numeric} for the hospitalisation risk of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation risks Default is 20\% hospitalisation (\code{0.2}) for the entire
population. If the \code{onset_to_hosp} argument is set to \code{NA} this argument
should also be set to \code{NA}. See details and examples for more information.}

\item{onset_to_death}{An \verb{<epidist>} object, an anonymous function for
the onset to death delay distribution, or \code{NA} to not simulate dates for
individuals that died.}

\item{onset_to_recovery}{An \verb{<epidist>} object, an anonymous function for
the onset to death delay distribution, or \code{NA} to not simulate dates for
individuals that recovered. Default is \code{NA} so by default cases that
recover get an \code{NA} in the \verb{$date_outcome} line list column.}

\item{hosp_death_risk}{Either a single \code{numeric} for the death risk for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death risks Default is 50\% death risk in hospitals
(\code{0.5}) for the entire population. If the \code{onset_to_death} argument is set
to \code{NA} this argument should also be set to \code{NA}. See details and examples
for more information.}

\item{non_hosp_death_risk}{Either a single \code{numeric} for the death risk for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death risks outside of hospitals. Default is 5\% death risk outside
of hospitals  (\code{0.05}) for the entire population. If the \code{onset_to_death}
argument is set to \code{NA} this argument should also be set to \code{NA}. See
details and examples for more information.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A \verb{<data.frame>} with one more column than input into \code{.data}.
Unless the column heading is already present in which the data is
overwritten.
}
\description{
The event date could be first contact, last contact or other.
}
\keyword{internal}
