% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_linelist.R
\name{.create_linelist}
\alias{.create_linelist}
\title{Create a mock line list without simulating to test internal functions that
require line list data}
\usage{
.create_linelist(scenario)
}
\arguments{
\item{scenario}{A \code{character} string with the state of the line list.
Options are:
\itemize{
\item "pre_date_last_contact"
\item "pre_date_first_contact"
\item "pre_hospitalisation"
\item "pre_deaths"
\item "pre_names"
\item "pre_ct"
}}
}
\value{
A \verb{<data.frame>}.
}
\description{
Create a mock line list without simulating to test internal functions that
require line list data
}
\details{
Each of the scenarios are created by taking a snapshot (i.e copy)
of the  line list \verb{<data.frame>} are each stage of the simulation
(\code{\link[=sim_linelist]{sim_linelist()}}) after setting the seed to 1 (\code{set.seed(1)}). The
scenarios are before each function call, for example,
\code{"pre_date_last_contact"} is the state of the line list prior to calling
\code{.add_date_contact(..., contact_type = "last")}, and \code{"pre_names"} is the
state of the line list prior to calling \code{\link[=.add_names]{.add_names()}}.
\subsection{Script to reproduce data:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{# load data required to simulate line list
serial_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# get onset to hospital admission from \{epiparameter\} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)

# get onset to death from \{epiparameter\} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
set.seed(1)

linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_rate = 0.5
)
}\if{html}{\out{</div>}}
}
}
\keyword{internal}
