% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_contacts_tbl.R
\name{.sim_contacts_tbl}
\alias{.sim_contacts_tbl}
\title{Create contacts table}
\usage{
.sim_contacts_tbl(.data, contact_tracing_status_probs)
}
\arguments{
\item{.data}{A \verb{<data.frame>} containing the infectious history from a
branching process simulation.}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}
}
\value{
A contacts \verb{<data.frame>}
}
\description{
Create contacts table
}
\keyword{internal}
