% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_internal.R
\name{.sim_internal}
\alias{.sim_internal}
\title{Internal simulation function called by the exported simulation functions
within \pkg{simulist}}
\usage{
.sim_internal(
  sim_type = c("linelist", "contacts", "outbreak"),
  contact_distribution,
  infect_period,
  prob_infect,
  onset_to_hosp = NULL,
  onset_to_death = NULL,
  hosp_risk = NULL,
  hosp_death_risk = NULL,
  non_hosp_death_risk = NULL,
  outbreak_start_date,
  add_names = NULL,
  add_ct = NULL,
  outbreak_size,
  population_age,
  case_type_probs = NULL,
  contact_tracing_status_probs = NULL,
  config
)
}
\arguments{
\item{sim_type}{A \code{character} string specifying which simulation function
this function is being called within.}

\item{contact_distribution}{An \verb{<epidist>} object or anonymous function for
the contact distribution. This is any discrete density function that
produces non-negative integers (including zero, \eqn{\mathbb{N}_0}) for the
number of contacts per infection.}

\item{infect_period}{An \verb{<epidist>} object or anonymous function for
the infectious period. This defines the duration from becoming infectious
to no longer infectious. In the simulation, individuals are assumed to
become infectious immediately after being infected (the latency period is
assumed to be zero). The time intervals between an infected individual and
their contacts are assumed to be uniformly distributed within the
infectious period.}

\item{prob_infect}{A single \code{numeric} for the probability of a secondary
contact being infected by an infected primary contact.}

\item{onset_to_hosp}{An \verb{<epidist>} object or anonymous function for
the onset to hospitalisation delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object or anonymous function for
the onset to death delay distribution.}

\item{hosp_risk}{Either a single \code{numeric} for the hospitalisation risk of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation risks Default is 20\% hospitalisation (\code{0.2}) for the entire
population. See details and examples for more information.}

\item{hosp_death_risk}{Either a single \code{numeric} for the death risk for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death risks Default is 50\% death risk in hospitals
(\code{0.5}) for the entire population. See details and examples for more
information.}

\item{non_hosp_death_risk}{Either a single \code{numeric} for the death risk for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death risks outside of hospitals. Default is 5\% death risk outside
of hospitals  (\code{0.05}) for the entire population. See details and examples
for more information.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{add_names}{A \code{logical} boolean for whether to add names to each row
of the line list. Default is \code{TRUE}.}

\item{add_ct}{A \code{logical} boolean for whether to add Ct values to each
confirmed case and \code{NA} otherwise for each case in the line list.
Default is \code{TRUE}. Ct refers to the Cycle threshold from a Real-time
PCR or quantitative PCR (qPCR).}

\item{outbreak_size}{A \code{numeric} vector of length 2 defining the minimum and
the maximum number of infected individuals for the simulated outbreak.
Default is \code{c(10, 1e4)}, so the minimum outbreak size is 10 infected
individuals, and the maximum outbreak size is 10,000 infected individuals.
Either number can be changed to increase or decrease the maximum or minimum
outbreak size to allow simulating larger or smaller outbreaks. If the
minimum outbreak size cannot be reached after running the simulation for
many iterations (internally) then the function errors, whereas if the
maximum outbreak size is exceeded the function returns the data early and a
warning stating how many cases and contacts are returned.}

\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{case_type_probs}{A named \code{numeric} vector with the probability of
each case type. The names of the vector must be \code{"suspected"}, \code{"probable"},
\code{"confirmed"}. Values of each case type must sum to one.}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A \verb{<data.frame>} if \code{sim_type} is \code{"linelist"} or \code{"contacts"}, or a
list of two \verb{<data.frame>}s if \code{sim_type} is \code{"outbreak"}.
}
\description{
This internal function simulates a line list, and
when \code{sim_type} is \code{"contacts"} or \code{"outbreak"} a contacts table as well.
}
\keyword{internal}
