% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_internal.R
\name{.sim_internal}
\alias{.sim_internal}
\title{Internal simulation function called by the exported simulation functions
within \pkg{simulist}}
\usage{
.sim_internal(
  sim_type = c("linelist", "contacts", "outbreak"),
  contact_distribution,
  contact_interval,
  prob_infect,
  onset_to_hosp,
  onset_to_death,
  hosp_risk,
  hosp_death_risk,
  non_hosp_death_risk,
  outbreak_start_date,
  add_names,
  add_ct,
  min_outbreak_size,
  population_age,
  case_type_probs,
  contact_tracing_status_probs,
  config
)
}
\arguments{
\item{sim_type}{A \code{character} string specifying which simulation function
this function is being called within.}

\item{contact_distribution}{An \verb{<epidist>} object or anonymous function for
the contact distribution. This is any discrete density function that
produces non-negative integers (including zero, \eqn{\mathbb{N}_0}) for the
number of contacts per infection.}

\item{contact_interval}{An \verb{<epidist>} object or anonymous function for
the contact interval. This is analogous to the serial interval or generation
time, but defines the time interval between an individual being
infected/infectious (in the simulation the latency period is assumed to be
zero) and having contact with another susceptible individual.}

\item{prob_infect}{A single \code{numeric} for the probability of a secondary
contact being infected by an infected primary contact.}

\item{onset_to_hosp}{An \verb{<epidist>} object or anonymous function for
the onset to hospitalisation delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object or anonymous function for
the onset to death delay distribution.}

\item{hosp_risk}{Either a single \code{numeric} for the hospitalisation risk of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation risks Default is 20\% hospitalisation (\code{0.2}) for the entire
population. See details and examples for more information.}

\item{hosp_death_risk}{Either a single \code{numeric} for the death risk for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death risks Default is 50\% death risk in hospitals
(\code{0.5}) for the entire population. See details and examples for more
information.}

\item{non_hosp_death_risk}{Either a single \code{numeric} for the death risk for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death risks outside of hospitals. Default is 5\% death risk outside
of hospitals  (\code{0.05}) for the entire population. See details and examples
for more information.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{add_names}{A \code{logical} boolean for whether to add names to each row
of the line list. Default is \code{TRUE}.}

\item{add_ct}{A \code{logical} boolean for whether to add Ct values to each
confirmed case and \code{NA} otherwise for each case in the line list.
Default is \code{TRUE}. Ct refers to the Cycle threshold from a Real-time
PCR or quantitative PCR (qPCR).}

\item{min_outbreak_size}{A single \code{numeric} defining the minimum chain size
for the simulated outbreak. Default is \code{10}. This can be increased when the
function should simulate a larger outbreak.}

\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{case_type_probs}{A named \code{numeric} vector with the probability of
each case type. The names of the vector must be \code{"suspected"}, \code{"probable"},
\code{"confirmed"}. Values of each case type must sum to one.}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A \verb{<data.frame>} if \code{sim_type} is \code{"linelist"} or \code{"contacts"}, or a
list of two \verb{<data.frame>}s if \code{sim_type} is \code{"outbreak"}.
}
\description{
This internal function simulates a line list, and
when \code{sim_type} is \code{"contacts"} or \code{"outbreak"} a contacts table.
}
\keyword{internal}
