% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_network_bp.R
\name{.sim_network_bp}
\alias{.sim_network_bp}
\title{Simulate a random network branching process model with a probability of
infection for each contact}
\usage{
.sim_network_bp(mean_contacts, contact_interval, prob_infect)
}
\arguments{
\item{mean_contacts}{A \code{numeric} for the mean number of contacts across the
population.}

\item{contact_interval}{An \verb{<epidist>} object or anonymous function for
the contact interval. The contact interval is similar to the generation time
or serial interval, but describes the time between becoming infected by a
contact and contacting another susceptible individual.}

\item{prob_infect}{A \code{numeric} for the probability of a secondary contact
being infected by an infected primary contact.}
}
\value{
A \verb{<data.frame>} with the contact and transmission chain data.
}
\description{
Simulate a branching process on a infinite network where the contact
distribution provides a function to sample the number of contacts of each
individual in the simulation. Each contact is then infected with the
probability of infection. The time between each infection is determined
by the contact interval function.
}
\details{
The contact distribution sampled takes the network effect
\eqn{q(n) ~ (n + 1)p(n + 1)} where \eqn{p(n)} is the probability density function
of a distribution, e.g., Poisson or Negative binomial. That is to say, the
probability of having choosing a contact at random by following up a
contact chooses individuals with a probability proportional to their number
of contacts. The plus one is because one of the contacts was "used" to
infect the person.
}
\keyword{internal}
