% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_contacts.R
\name{sim_contacts}
\alias{sim_contacts}
\title{Simulate contacts for an infectious disease outbreak}
\usage{
sim_contacts(
  R,
  serial_interval,
  contact_distribution,
  outbreak_start_date = as.Date("2023-01-01"),
  min_outbreak_size = 10,
  population_age = c(1, 90),
  contact_tracing_status_probs = c(under_followup = 0.7, lost_to_followup = 0.2, unknown
    = 0.1),
  config = create_config()
)
}
\arguments{
\item{R}{A single \code{numeric} for the reproduction number.}

\item{serial_interval}{An \verb{<epidist>} object or anonymous function for
the serial interval.}

\item{contact_distribution}{An \verb{<epidist>} object or anonymous function for
the contact distribution. This is similar to the offspring distribution,
but describes the heterogeneity of all contact patterns, including contacts
that go on to be infected or not infected.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{min_outbreak_size}{A single \code{numeric} defining the minimum chain size
for the simulated outbreak. Default is \code{10}. This can be increased when the
function should simulate a larger outbreak.}

\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A contacts \verb{<data.frame>}
}
\description{
Simulate contacts for an infectious disease outbreak
}
\examples{
# load data required to simulate contacts
serial_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

contact_distribution <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "contact_distribution",
  prob_distribution = "pois",
  prob_distribution_params = c(l = 5)
)

contacts <- sim_contacts(
  R = 1.1,
  serial_interval = serial_interval,
  contact_distribution = contact_distribution
)
}
\author{
Joshua W. Lambert, Carmen Tamayo
}
