% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_linelist.R
\name{sim_linelist}
\alias{sim_linelist}
\title{Simulate a linelist}
\usage{
sim_linelist(
  R,
  serial_interval,
  onset_to_hosp,
  onset_to_death,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  add_ct = FALSE,
  min_chain_size = 10,
  age_range = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  hosp_rate = list(),
  hosp_to_death_rate = list(),
  include_contacts = FALSE,
  ...
)
}
\arguments{
\item{R}{A single \code{numeric} for the reproduction number.}

\item{serial_interval}{An \verb{<epidist>} object for the serial interval.}

\item{onset_to_hosp}{An \verb{<epidist>} object for the onset to hospitalisation
delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object for the onset to death delay
distribution.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{add_names}{A \code{logical} boolean for whether to add names to each row
of the linelist. Default is \code{TRUE}.}

\item{add_ct}{A \code{logical} boolean for whether to add Ct values to each row
of the linelist. Default is \code{FALSE}.}

\item{min_chain_size}{A single \code{numeric} defining the minimum chain size for
the simulated outbreak. Default is \code{10}. This can be increased when the
function should simulate a larger outbreak.}

\item{age_range}{A \code{numeric} vector with two elements. The first is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]).}

\item{case_type_probs}{A named \code{numeric} vector with the probability of
each case type. The names of the vector must be \code{"suspected"}, \code{"probable"},
\code{"confirmed"}. Values of each case type must sum to one.}

\item{hosp_rate}{A \code{list} for the age-stratified hospitalisation rate.
\strong{Currently not implemented}.}

\item{hosp_to_death_rate}{A \code{list} for the age-stratified rate at which
those admitted to hospital go onto die. \strong{Currently not implemented}.}

\item{include_contacts}{A \code{logical} boolean to determine whether contacts
are added to the linelist. \strong{Currently not implmented}.}

\item{...}{\link{dots} Extra arguments to be passed to other functions.
\strong{Currently not used and will return a warning if extra arguments are
supplied}.}
}
\value{
A linelist \verb{<data.frame>}
}
\description{
The linelist is simulated using a branching process and
parameterised with previously published epidemiological parameters.
}
