% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_linelist.R
\name{sim_linelist}
\alias{sim_linelist}
\title{Simulate a line list}
\usage{
sim_linelist(
  contact_distribution,
  infect_period,
  prob_infect,
  onset_to_hosp,
  onset_to_death,
  onset_to_recovery = NA,
  hosp_risk = 0.2,
  hosp_death_risk = 0.5,
  non_hosp_death_risk = 0.05,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  add_ct = TRUE,
  outbreak_size = c(10, 10000),
  population_age = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  config = create_config()
)
}
\arguments{
\item{contact_distribution}{An \verb{<epidist>} object or anonymous function for
the contact distribution. This is any discrete density function that
produces non-negative integers (including zero, \eqn{\mathbb{N}_0}) for the
number of contacts per infection.}

\item{infect_period}{An \verb{<epidist>} object or anonymous function for
the infectious period. This defines the duration from becoming infectious
to no longer infectious. In the simulation, individuals are assumed to
become infectious immediately after being infected (the latency period is
assumed to be zero). The time intervals between an infected individual and
their contacts are assumed to be uniformly distributed within the
infectious period.}

\item{prob_infect}{A single \code{numeric} for the probability of a secondary
contact being infected by an infected primary contact.}

\item{onset_to_hosp}{An \verb{<epidist>} object, an anonymous function for
the onset to hospitalisation delay distribution, or \code{NA} to not simulate
hospitalisation (admission) dates.}

\item{onset_to_death}{An \verb{<epidist>} object, an anonymous function for
the onset to death delay distribution, or \code{NA} to not simulate dates for
individuals that died.}

\item{onset_to_recovery}{An \verb{<epidist>} object, an anonymous function for
the onset to death delay distribution, or \code{NA} to not simulate dates for
individuals that recovered. Default is \code{NA} so by default cases that
recover get an \code{NA} in the \verb{$date_outcome} line list column.}

\item{hosp_risk}{Either a single \code{numeric} for the hospitalisation risk of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation risks Default is 20\% hospitalisation (\code{0.2}) for the entire
population. If the \code{onset_to_hosp} argument is set to \code{NA} this argument
should also be set to \code{NA}. See details and examples for more information.}

\item{hosp_death_risk}{Either a single \code{numeric} for the death risk for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death risks Default is 50\% death risk in hospitals
(\code{0.5}) for the entire population. If the \code{onset_to_death} argument is set
to \code{NA} this argument should also be set to \code{NA}. See details and examples
for more information.}

\item{non_hosp_death_risk}{Either a single \code{numeric} for the death risk for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death risks outside of hospitals. Default is 5\% death risk outside
of hospitals  (\code{0.05}) for the entire population. If the \code{onset_to_death}
argument is set to \code{NA} this argument should also be set to \code{NA}. See
details and examples for more information.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{add_names}{A \code{logical} boolean for whether to add names to each row
of the line list. Default is \code{TRUE}.}

\item{add_ct}{A \code{logical} boolean for whether to add Ct values to each
confirmed case and \code{NA} otherwise for each case in the line list.
Default is \code{TRUE}. Ct refers to the Cycle threshold from a Real-time
PCR or quantitative PCR (qPCR).}

\item{outbreak_size}{A \code{numeric} vector of length 2 defining the minimum and
the maximum number of infected individuals for the simulated outbreak.
Default is \code{c(10, 1e4)}, so the minimum outbreak size is 10 infected
individuals, and the maximum outbreak size is 10,000 infected individuals.
Either number can be changed to increase or decrease the maximum or minimum
outbreak size to allow simulating larger or smaller outbreaks. If the
minimum outbreak size cannot be reached after running the simulation for
many iterations (internally) then the function errors, whereas if the
maximum outbreak size is exceeded the function returns the data early and a
warning stating how many cases and contacts are returned.}

\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{case_type_probs}{A named \code{numeric} vector with the probability of
each case type. The names of the vector must be \code{"suspected"}, \code{"probable"},
\code{"confirmed"}. Values of each case type must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A line list \verb{<data.frame>}
}
\description{
The line list is simulated using a branching process and
parameterised with user defined epidemiological parameters.
}
\details{
For age-stratified hospitalised and death risks a \verb{<data.frame>}
will need to be passed to the \code{hosp_risk} and/or \code{hosp_death_risk}
arguments. This \verb{<data.frame>} should have two columns:
\itemize{
\item \code{age_limit}: a column with one \code{numeric} per cell for the lower bound
(minimum) age of the age group (inclusive).
\item \code{risk}: a column with one \code{numeric} per cell for the proportion
(or probability) of hospitalisation for that age group. Should be between
0 and 1.
}

For an age structured population, a \verb{<data.frame>} with two columns:
\itemize{
\item \code{age_range}: a column with characters specifying the lower and upper bound
of that age group, separated by a hyphen (-). Both bounds are inclusive
(integers). For example, an age group of one to ten would be given as
\code{"1-10"}.
\item \code{proportion}: a column with the proportion of the population that are in
that age group. Proportions must sum to one.
}
}
\examples{
# load data required to simulate line list
contact_distribution <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "contact distribution",
  prob_distribution = "pois",
  prob_distribution_params = c(mean = 2)
)

infect_period <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "infectious period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
# example with single hospitalisation risk for entire population
linelist <- sim_linelist(
  contact_distribution = contact_distribution,
  infect_period = infect_period,
  prob_infect = 0.5,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = 0.5
)
head(linelist)

# example with age-stratified hospitalisation risk
# 20\% for over 80s
# 10\% for under 5s
# 5\% for the rest
age_dep_hosp_risk <- data.frame(
  age_limit = c(1, 5, 80),
  risk = c(0.1, 0.05, 0.2)
)
linelist <- sim_linelist(
  contact_distribution = contact_distribution,
  infect_period = infect_period,
  prob_infect = 0.5,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = age_dep_hosp_risk
)
head(linelist)
}
\author{
Joshua W. Lambert, Carmen Tamayo
}
