% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outbreak.R
\name{sim_outbreak}
\alias{sim_outbreak}
\title{Simulate a line list and a contacts table}
\usage{
sim_outbreak(
  contact_distribution,
  contact_interval,
  prob_infect,
  onset_to_hosp,
  onset_to_death,
  hosp_risk = 0.2,
  hosp_death_risk = 0.5,
  non_hosp_death_risk = 0.05,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  add_ct = TRUE,
  min_outbreak_size = 10,
  population_age = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  contact_tracing_status_probs = c(under_followup = 0.7, lost_to_followup = 0.2, unknown
    = 0.1),
  config = create_config()
)
}
\arguments{
\item{contact_distribution}{An \verb{<epidist>} object or anonymous function for
the contact distribution. This is any discrete density function that
produces non-negative integers (including zero, \eqn{\mathbb{N}_0}) for the
number of contacts per infection.}

\item{contact_interval}{An \verb{<epidist>} object or anonymous function for
the contact interval. This is analogous to the serial interval or generation
time, but defines the time interval between an individual being
infected/infectious (in the simulation the latency period is assumed to be
zero) and having contact with another susceptible individual.}

\item{prob_infect}{A single \code{numeric} for the probability of a secondary
contact being infected by an infected primary contact.}

\item{onset_to_hosp}{An \verb{<epidist>} object or anonymous function for
the onset to hospitalisation delay distribution.}

\item{onset_to_death}{An \verb{<epidist>} object or anonymous function for
the onset to death delay distribution.}

\item{hosp_risk}{Either a single \code{numeric} for the hospitalisation risk of
everyone in the population, or a \verb{<data.frame>} with age specific
hospitalisation risks Default is 20\% hospitalisation (\code{0.2}) for the entire
population. See details and examples for more information.}

\item{hosp_death_risk}{Either a single \code{numeric} for the death risk for
hospitalised individuals across the population, or a \verb{<data.frame>} with age
specific hospitalised death risks Default is 50\% death risk in hospitals
(\code{0.5}) for the entire population. See details and examples for more
information.}

\item{non_hosp_death_risk}{Either a single \code{numeric} for the death risk for
outside of hospitals across the population, or a \verb{<data.frame>} with age
specific death risks outside of hospitals. Default is 5\% death risk outside
of hospitals  (\code{0.05}) for the entire population. See details and examples
for more information.}

\item{outbreak_start_date}{A \code{date} for the start of the outbreak.}

\item{add_names}{A \code{logical} boolean for whether to add names to each row
of the line list. Default is \code{TRUE}.}

\item{add_ct}{A \code{logical} boolean for whether to add Ct values to each row
of the line list. Default is \code{TRUE}.}

\item{min_outbreak_size}{A single \code{numeric} defining the minimum chain size
for the simulated outbreak. Default is \code{10}. This can be increased when the
function should simulate a larger outbreak.}

\item{population_age}{Either a \code{numeric} vector with two elements or a
\verb{<data.frame>} with age structure in the population. Use a \code{numeric} vector
to specific the age range of the population, the first element is the lower
bound for the age range, and and the second is the upper bound for the age
range (both inclusive, i.e. [lower, upper]). The \verb{<data.frame>} with
age groups and the proportion of the population in that group. See details
and examples for more information.}

\item{case_type_probs}{A named \code{numeric} vector with the probability of
each case type. The names of the vector must be \code{"suspected"}, \code{"probable"},
\code{"confirmed"}. Values of each case type must sum to one.}

\item{contact_tracing_status_probs}{A named \code{numeric} vector with the
probability of each contact tracing status. The names of the vector must
be \code{"under_followup"}, \code{"lost_to_followup"}, \code{"unknown"}. Values of each
contact tracing status must sum to one.}

\item{config}{A list of settings to adjust the randomly sampled delays and
Ct values (if \code{add_ct = TRUE}). See \code{\link[=create_config]{create_config()}} for more information.}
}
\value{
A list with two elements:
\enumerate{
\item A line list \verb{<data.frame>}
\item A contacts \verb{<data.frame>}
}
}
\description{
The line list and contacts are simulated using a branching
process and parameterised with previously published epidemiological
parameters.
}
\details{
For age-stratified hospitalised and death risks a \verb{<data.frame>}
will need to be passed to the \code{hosp_risk} and/or \code{hosp_death_risk}
arguments. This \verb{<data.frame>} should have two columns:
\itemize{
\item \code{age_limit}: a column with one \code{numeric} per cell for the lower bound
(minimum) age of the age group (inclusive).
\item \code{risk}: a column with one \code{numeric} per cell for the proportion
(or probability) of hospitalisation for that age group. Should be between
0 and 1.
}

For an age structured population, a \verb{<data.frame>} with two columns:
\itemize{
\item \code{age_range}: a column with characters specifying the lower and upper bound
of that age group, separated by a hyphen (-). Both bounds are inclusive
(integers). For example, an age group of one to ten would be given as
\code{"1-10"}.
\item \code{proportion}: a column with the proportion of the population that are in
that age group. Proportions must sum to one.
}
}
\examples{
# load data required to simulate outbreak data
contact_distribution <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "contact distribution",
  prob_distribution = "pois",
  prob_distribution_params = c(mean = 2)
)

contact_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)

outbreak <- sim_outbreak(
  contact_distribution = contact_distribution,
  contact_interval = contact_interval,
  prob_infect = 0.5,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death
)
}
\author{
Joshua W. Lambert
}
