% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulist-package.R
\docType{package}
\name{simulist-package}
\alias{simulist}
\alias{simulist-package}
\title{simulist: Simulate Disease Outbreak Line List and Contacts Data}
\description{
Tools to simulate realistic raw case data for an epidemic in the form of line lists and contacts using a branching process. Simulated outbreaks are parameterised with epidemiological parameters and can have age-structured populations, age-stratified hospitalisation and death risk and time-varying case fatality risk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/simulist}
  \item \url{https://epiverse-trace.github.io/simulist/}
  \item Report bugs at \url{https://github.com/epiverse-trace/simulist/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Carmen Tamayo \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor, reviewer]
  \item Pratik R. Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [contributor, reviewer]
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [reviewer]
  \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [reviewer]
  \item London School of Hygiene and Tropical Medicine, LSHTM (00a0jsq62) [copyright holder]
}

}
\keyword{internal}
